Circumvent a very annoying compiler warning
"warning: ISO C++ says that these are ambiguous, even though the worst conversion for the first is better than the worst conversion for the second:"

Unravelling this is too much work when there's an alternate solution.
Besides, the intent is to append the string in-place,
so let's just do that.