Layer2 registry in Layer3 needs to use shared pointers
otherwise Layer2 objects are removed as soon as there's no shared_ptr to
them, which means that configured Layer2 drivers are closed immediately,
which is not what we want. At all. :-P

This means that de-registering the L2 driver can't happen in the destructor.
Therefore a RunDone method is added which does that as soon as the loop
terminates.

TODO: do this for L2 drivers which don't have their own thread.