Group cache must be initialized to work correctly
Due to some code changes the group cache didn't work anymore. In the
BCU SDK it was special cased, now it's implemented as yet another Layer2
interface. However that means it has to be handled like an L2 interface,
especially the init() function must be called such that it can register
with the L3 object and participate in all messaging.

Fix #91

Signed-off-by: Thomas Dallmair <dallmair@users.noreply.github.com>