Fix routing counter=7 on packets sent by knxtool (issue #45)
Previously, the default hop count of every L_Data_PDU was 7, so that value was
used for every packet originating in clients like knxtool. Change this to only
set the value 7 explicitly for requests coming in through T_Broadcast, and set
the default to 6 instead (used for T_Group, T_TPDU, T_Individual, T_Connection,
and GroupSocket).

Signed-off-by: Thomas Dallmair <dallmair@users.noreply.github.com>