Fix memory leak on client requests
The client request is mapped to the requested handler (e.g. A_Group), which
creates a layer 4 object (e.g. T_Group) in its constructor and registers that
one with the connection's layer 3 object. But the l3 object lives as long as the
process and keeps all the l2's around all the time.

Solution is to unregister the layer 2 on deconstruction of the request handler.

Signed-off-by: Thomas Dallmair <dallmair@users.noreply.github.com>