Add --single-port/--multi-port options to control how to use multicast datagrams.
The single-port mode uses one port (3671) for both advertising and datagram
transmission. Multi-port mode splits that and advertises a random port for
datagrams.

The advantage of multi-port mode is that you can run more than one knxd on
a system; this is useful when a driver is unstable and you want to
selectively restart it (knxd doesn't (yet?) support dynamically attaching
new drivers).

ETS5 currently ignores the advertised port and thus cannot use a knxd
running in multi-port mode. Therefore, single-port mode is the default.