Fix linkage to filters and driver.
The router holds shared pointers to the LinkConnect object,
which in turn holds pointers to the first filter and to the driver.
Each filter holds a pointer to the next filter (or to the driver).

The router is deemed static, thus it can be referenced directly.
All other back links shall be done with weak pointers.

Also, Thou Shalt Not Call Shared_From_This() From A Constructor.