Add a "remap" filter.
This is the inverse of "single"; it fixes remote addresses to always show
the link address assigned by knxd.

Use this filter for obstinate remote code which has its own idea of what
its address should be, instead of using the one knxd assigns to it.

This filter can also be used to connect remote networks with colliding phys
addresses.