Fix `--Interface` not listening to specified interface
Calling IP_ADD_MEMBERSHIP with the ip_mreq struct and imr_address set to
INADDR_ANY (0) results in the kernel deciding[0] which interface to use
based on which interface has the default route and not any interface as
one might expect.

This results in the `--Interface` option being nonfunctional unless the
chosen interface also happens to be the one containing the default
route. Multicast messages are sent to the correct interface as set by
IP_MULTICAST_IF but everything sent to us is thrown away by the kernel
since we're listening to a different interface.

To fix this, use the ip_mreqn struct instead and pass the interface
number.

[0] https://github.com/torvalds/linux/blob/87adedeba51a822533649b143232418b9e26d08b/net/ipv4/igmp.c#L1827