Ensure that the buffer will not be overrun
With `-Wformat-truncation=`, GCC does a worst case calculation for
format strings. In this, %d can need up to 11 characters (10 digits and
a leading minus), which would lead to a buffer overrun of `logfile`.

As this won't happen in practice, it does not make sense to increase
the buffer size. Instead, update the format string to make it clear that
the written string is shorter.